feat: refactor authenticator
