--- crates/authenticator/src/api_types.rs
@@ -1,1 +1,0 @@
-pub use world_id_primitives::api_types::*;
